Compare vote totals

Compare vote totals with a reference dataset. By default, the reference totals are the state_totals dataset from Dave Leip's Atlas of U.S.Presidential Elections, for candidate values of Donald Trump, Hillary Clinton, Gary Johnson, and "Total" (the sum of votes for these three and all other candidates).

compare_votes(.data, reference = NULL, by = c("state_postal", "office",
  "candidate"), all.y = FALSE)

Arguments

.data
Precinct returns dataframe.

reference
A dataframe of reference vote totals. See details.

by
A vector of shared column names in .data and reference to merge on.

all.y
Perform a full join, instead of a left join. This will include in the result all rows from the reference dataset.

Value

A copy of .data with state-candidate vote totals and their difference from the reference totals.

Details

We sum votes in each dataframe by the variables specified in argument by, and then left-join .data and the reference dataframe on the same variables.

Examples

# Use default reference totals from https://uselectionatlas.org
.data = data.frame(office = 'US President', candidate = c('Gary Johnson',
  'Donald Trump'), state_postal = 'AL', votes = c(44000, 1318200))
compare_votes(.data)

#>    state_postal       office    candidate   votes votes_reference
#> 1:           AL US President Donald Trump 1318200         1318255
#> 2:           AL US President Gary Johnson   44000           44467
#>    diff_reference
#> 1:            -55
#> 2:           -467

# Use an arbitrary dataframe for reference totals
reference = data.frame(office = 'US President', candidate = c('Gary Johnson',
  'Donald Trump'), state_postal = 'AL', votes = c(43022, 1327200))
compare_votes(.data, reference)

#>    state_postal       office    candidate   votes votes_reference
#> 1:           AL US President Donald Trump 1318200         1327200
#> 2:           AL US President Gary Johnson   44000           43022
#>    diff_reference
#> 1:          -9000
#> 2:            978